+++ app/src/main/java/com/example/annabujak/weather4runners/Database/ChosenHour.java
@@ -7,7 +7,7 @@ import java.util.Date;
  */
 
 public class ChosenHour {
-    int Id;
+    long Id;
     Date Date;
     int Hour;
 
@@ -25,10 +25,10 @@ public class ChosenHour {
     public int getHour(){
         return Hour;
     }
-    public void setId(int id){
+    public void setId(long id){
         Id = id;
     }
-    public int getId(){
+    public long getId(){
         return Id;
     }
 }
